Maze generation is mostly done.
Comments:
The maze seems to be generating properly now, through all passes. Still
need to make sure all entrances are aligned and place a different tile
when a square has all 4 walls.

Changes:
Added a pass to the maze generation that clears any tiles an entity is
touching (so they don't get stuck in walls).

Updated the tile map to align the entrance.