Build move improvements.
Changes:
Added MainOverlay with hint text for build mode.

MainScreen now tracks whether the player is in the build area or
not and restricts/closes build mode.

Added a client signal for playerPositionChanged.

Added server logic to validate that the player is in the build area
before applying tile updates.

Added a Shared module. It was always planned to be there, but this is
the first time we've needed it. Unfortunately it only has a header
right now, so we can't build a binary from it. The cmake logic for
making it a static library target is included and commented out, though.