Improved build/package process.
Changes:
Changed default serverIP in the client's UserConfig.json to be
localhost, so it works when people try it out locally.

On Windows, the Client and Server builds now copy the SDL2 DLLs into
the build folder, since they aren't installed system-wide.

Added an install step to Client/Server that packages the app,
resources, and (on Windows) DLLs. Going forward, we'll use this
instead of a package script.

Updated README.