Updated build area extent, default to layer 1.
Changes:
Updated the build area extent so the client is able to open
build mode while being 1 tile outside of the placeable area.
This makes it so people can't wall off the build area and keep
people from being able to take the walls down.

Changed the default build mode layer from 0 to 1. People keep
accidentally replacing the floor and get confused. This is more
friendly.