Map update, bugfixes.
Changes:
Updated the tile map.

Fixed build mode showing "Layer 1" on first open, but actually
acting on layer 0.

Fixed a pop and swap bug in MazeGenerationSystem. If the element
to remove is the last element, you need to set the iterator to
end() afterwards because pop_back() invalidates iterators after
the removed element.

Added an assert for safety.

Expanded the build mode area west by 1 tile.