System bugfixes.
Changes:
Fixed a bug in MazeGenerationSystem caused by swap and popping at
the end of a vector (swapping invalidates all iterators after the
swapped element. in this case, the end iterator).
We could've fixed it by just adding an extra condition, but
we instead switched to std::erase_if because it's much cleaner and
completely avoids the problem.

Fixed a bug in PlantSystem caused by using a too-large random
range as an index. I fixed this previously just below, but failed
to see this one.