Stopped console opening on windows, added app icon
Changes:
Added flag to make Client and Server into GUI applications on
Windows instead of console applications. This gets rid of the
console that was opening up in the background.

Added .rc and .ico, along with relevant CMake. Now the app has an
icon on windows, and the icon scales to larger sizes when used
in shortcuts and such.