Improved Wall tile tool.
Changes:
Updated SpriteData.json to include sprite sets for all of the
sprites.

Wall tool updates to match the engine improvements around NW gap
fills (updating the sprite set type when the surrounding walls are
changes).

Fixed an issue where WallTool was pushing TileAddLayers for both
the West wall and NorthEastGapFill when adding a west wall to a
tile with a North wall (the map does the NE automatically, we just
need to push the West).