Removed default entity interactions.
Changes:
Removed default entity interactions. We need the left click
for targeting, so right click will be the only way to
interact. This should be fine--it'd be annoying to players if
we used auto-movement since it'd make it harder to walk around,
but we use direct control so this is only forcing them to do an
extra click to interact with entities. Items remain the same.