Updated permissions checks.
Changes:
We now have multiple build mode areas, each of which has the same
permissions.

Added protection for our NPCs and items.
The behavior around trying to do entity or item stuff in the entrance
to an area is non-ideal (it'll restrict you without feedback), and
the same for trying to edit a protected entity. Hopefully this is
obvious enough, if not we can re-visit.